+++ lib/generators/sorcery/templates/migration/core.rb
@@ -1,12 +1,13 @@
 class SorceryCore < ActiveRecord::Migration
   def change
     create_table :<%= model_class_name.tableize %> do |t|
-      t.string :username,         :null => false  # if you use another field as a username, for example email, you can safely remove this field.
-      t.string :email,            :default => nil # if you use this field as a username, you might want to make it :null => false.
-      t.string :crypted_password, :default => nil
-      t.string :salt,             :default => nil
+      t.string :email, null: false
+      t.string :crypted_password
+      t.string :salt
 
       t.timestamps
     end
+
+    add_index :<%= model_class_name.tableize %>, :email, unique: true
   end
 end