+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## 0.8.3 (not released)
 
+* Don't add `:username` field for User.
+* Config option `username_attribute_names` is now `:email` by default instead of `:username`.
+
+  If you're using `username` as main field for users to login, you'll need to tune your Sorcery config:
+
+    ```ruby
+    config.user_config do |user|
+      # ...
+      user.username_attribute_names = [:username]
+    end
+    ```
 
 ## 0.8.2
 <!-- TO BE WRITTEN -->